Fix direct_install linker and make it the default.
DirectInstallLinker is a minor tweak to CrossSiteIFrameLinker
that makes it easier to read and debug the JavaScript.
The .js file directly contains the JavaScript instead of encoding
it as a list of strings to be installed via a function call.

DirectInstallLinker was unused (and actually broken), but
Super Dev Mode uses the same approach.

Bug: issue 8578
Change-Id: I4651ac7a2aeb41a51b60fe55d69d8f47a1119a83