Fix HtmlUnit exception handling in web mode tests.
To get exceptions from JSNI calls in HtmlUnit dev mode, we need to
configure HtmlUnit to throw exceptions on script error. However this
configuration also causes HtmlUnit to exit in response to any
uncaught JavaScript exception. As a result, this causes mysterious
timeouts in HtmlUnit tests.

As this option (throwExceptionsOnScriptError) is not needed for web
mode tests, this patch changes the run style to enable the option
only for dev mode tests so that we will not experience such timeouts
in HtmlUnit web mode tests (though dev-mode is still affected by the
problem).

Bug: issue 8007
Change-Id: I3f649d1a3cc01aff6fcf31741c0e2781cccc9b36
Review-Link: https://gwt-review.googlesource.com/#/c/6613/