Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

It's not clear to me what the problem is with the current setup. It's the standard Jetty setup so I don't understand why it would fail.

IntrospectorCleaner for example is not a "server class" so it'll be loaded from the WebAppClassLoader, falling back to its parent classloader, so there's no reason that it would fail here AFAICT.

(note: Jetty will, by default, create a WebAppClassLoader whose parent is the context class loader, or the classloader of the WebAppClassLoader class or if all that fails then the system classloader)

Also, note that outside Google, Jetty is not repackaged (no c.g.g.thirdparty prefix)

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1