Distinguish between Java arrays and native arrays for devirtualizing.
Dispatchs on arrays need to be different for Java arrays, which are
dispatched to Object methods and native JavaScript arrays (such as
those created in JSNI, which need to be dispatched to
JavaScriptObject methods instead.

This patch restores the behaviour of arrays prior to the patch
devirtualizing array method calls.

Change-Id: I8053f2b1bbade6321911d66932f85ea58c8e0c42