Update patch set 1
Patch Set 1:

Thanks Thomas.

As I digged through the source code of jetty, now I understand the classloader tricks in Jetty better and what isSystemClass, isServerClass etc means.

What was happening is, there is a rule in WebAppClassLoader to whitelist and blacklist server classes[1] and jarjar was matching and changing the rules for the positive match but not negative matches as they contain '-' in the string. This was causing the rebased servlet listener classes to be blacklisted and not loaded by WebAppClassLoader.

However as the jarjar doesn't accept '-' in patterns, I'll need to patch our copy of jarjar, if that works well we will be able get away by changing the jarjar rules.

[1] http://grepcode.com/file/repo1.maven.org/maven2/org.eclipse.jetty.aggregate/jetty-all/8.1.11.v20130520/org/eclipse/jetty/webapp/WebAppContext.java#132

Patch-set: 1