Update patch set 12
Patch Set 12:

I'm not sure if it is this commit or another one, but I have rebased my java 8 branch on top of this and found super dev mode blew up.

var $intern_0 = {3:1, 4:1}, $intern_1 = {3:1, 5:1};
var jl = {},  = {}, cggcc = {}, cggcci = {}, 

Note the , = {}
This is caused by classes in the default package resolving to a scope name of "", which is clearly a syntax error.  I encountered this because generated lambdas go into "", so as a temporary workaround I am putting lambdas in "lambda." instead.  

So, if this change is due to this commit, we should fix it here, otherwise this would be a bug in a merged commit somewhere.  If the fix should go elsewhere, I'd be glad to take it on if someone can point me to the place where I would need to make the package-name change to shim "".  If not, I could make a test for default package names in draftCompiles and contribute that instead.  Although we could probably write the test and run a bisectability script to find the commit responsible, I figured it would be easier to just let you guys know since your working knowledge here is far greater than mine.

Patch-set: 12
Label: Verified=0