Update patch set 1
Patch Set 1:

> I did it that was first :) I've just wondered if we've been
 > overeager to make our own emulations in the past, so I figured I'd
 > check with you in case you wanted it or you had another solution.

Yeah, historically GWT has chosen not to include any emulation that are inherently not compatible with client code or useful in client context (e.g. io, reflection and their friends). I recently go over the gwt/corp/emul and see if we want to move anything there but I didn't see on that would fit into this scope.

I don't know if we are going to change that decision anytime soon and that's definitely not something that we are discussing right now.
Perhaps it may change in the future and when that happens somebody from GWT team will need to go over gwt/corp/emul and move stuff that would be useful in GWT proper.

Patch-set: 1