Upgrade the GWT locale data to CLDR 24
(This is an export of a change already committed at Google.
Please do review it, but for ease of merging I'd like to commit
it as-is and then fix any errors found in a separate patch.)

Original description below.

This change list updates locale data, something not done since 2010.
So the changes are more extensive than it should normally be. One of
the important one was the the negative format for English changing
from (123) to -123. Big decision, debated in the CLDR committee quite a
bit. So that one "shacked the boat" a bit. But it is consistent with
what Google is doing across products, with the big shared libraries
(ICU for C/C++/Java, Closure for Javascript).

Change-Id: Iba5e039aaf5085f62a0265e7c6ebf485b261e206