Update patch set 1
Patch Set 1:

I'll look at it tonight.

Do you have a diff for Google-specific changes to the CLDR files?  I'm not sure where it goes now, but previously we checked in updates to gwt-tools, so they could be reproduced externally (plus the tools are there to regenerate these files).

Patch-set: 1