Found when testing SuperDevMode + incremental compilation.
In LibraryCompiler, makes library closing more robust, reloads the
written library as a ready only library when needing access to its
contained permutation result, and locks down more options in
normalizeOptions().

Switches all Condition toSource generation to JSNI so that it's also
possible to switch classLiteral == classLiteral comparison to be based
on JSNI class literal references, to get around private class
restrictions. Cascades to cause PropertyProviderRegistratorGenerator and
RuntimeRebindRuleGenerator to output JSNI functions.

Implements encoding/decoding of all ZipLibrary string multimap values so
that it is no longer necessary to throw an error if some content tries
to use one of the intended delimiter. Consequently, also revs
ZipLibraries version number.

LibraryGroup now actively prevents duplicate source inclusion to prevent
weird runtime results.

Adds a guard in CompilationUnitTypeOracleUpdater to not process classes
it has previously processed (to match the same expectation in
TypeOracle).

Added the missing "self" entry in
JTypeOracle.getSuperHierarchyTypes(type), that was resulting in
too-small castmaps in separate compilation.

Moved gwtOnLoad() output in GenerateJavaScriptAST to after class literal
insertion, so that bootstrap execution code will not error when
referring to class literals.

Added a very clear error in GwtAstBuilder to warn users who have invoked
a library compile but are missing a dependency on
com.google.gwt.user.User and thus are missing some Enum infrastructure.

Updates GWT.create() verification in UnifyAst and ReboundTypeRecorder to
not throw errors when seeing GWT.create() calls that don't take a
classLiteral parameter but ONLY IF it is occurring inside of the
GWT.java class itself.

Made ResolveRuntimeTypeReferences.java library compilation compatible by
ensuring that runtime type references in annotations and in the
ClassLiteralHolder class *are* seen.

Fixed a bug where UnifyAst.assimilateLibraryUnit() was calling and
operating on the results of referencedCompilationUnit.getTypes() *twice*
which actually returns a different copy each time.

Change-Id: Ia8062cc77ea3cd4cd5761b8251f875df1aafed96