Update patch set 6
Patch Set 6:

Hmm, this is a predecessor of what I was working on and a bit dated so I don't remember the exact state it was in.  I haven't worked on it in about 6 months, and I really need to get back to it, though a lot of the reason for doing some things the way they were is to improve DevMode performance, and it sounds like that may not be an issue for long.

So feel free to look at it as it has the basic approach and APIs, but the implementation details have changed a bit and I won't have any time to work on it until after my taxes are done.

Patch-set: 6