Update patch set 1
Patch Set 1:

These are good questions but, despite the author field, I'm not the author of this patch; another Googler actually did the work.

The author wants to start a conversation about i18n in GWT in general but he's not comfortable doing it publicly, so I'll start a separate email thread.

Patch-set: 1
Label: Verified=0