Fix JsStackEmulator bug for function calls
Evaluating expressions for function call parameters may
change the location to a different file due to inlining.
We need to set the location back before calling the function,
or the wrong location will be on the stack when creating
a stack trace.

(This can happen when a method call to create an exception's
message has been inlined.)

Change-Id: I84106acff5c089fc75e0cf99516f15a14a54886b