Update patch set 6
Patch Set 6:

Apart from the performance in dev-mode, the fact that the i18n api works in JVM is very interesting for junit testing of activities/presenters, and the usage of it in server side being coherent with client, etc.

So if this patch is good enought and it does not break anything important, could we approve it?, and then change the implementation when you finish your current work? This way could be easier to following incoming changes.

Patch-set: 6