Update patch set 1
Patch Set 1:

> It is actually there reference class so if you say @Foo:bar(*) you can access 
> all methods of Foo, public/protected methods of SuperFoo and package
> private methods of SuperFoo if Foo and SuperFoo are in the same package.
>
> It does not matter where the reference appears say 
> class SubFoo extends Foo {
> .. @Foo::bar();
> }

To be honest I didn't fully understand your response wrt my comment so I'll reply by assuming you misunderstood me :)

We should apply java visibility semantics + relax it for the private members of the target class.
If that's difficult to achieve, the alternative is to enforce "public" visibility for all members declared in the parents.

So if we follow option 1, calling SuperFoo::bar from SomeRandomClass via Foo:bar should work if and only if SomeRandomClass is visible from java perspective. IIUC, your implementation uses relation of Foo & SuperFoo to determine if it is callable which sounds incorrect to me.

Patch-set: 1
Label: Code-Review=0