Fix JsStackEmulator for constructor and function calls
Evaluating the arguments to a constructor or function call
may change the location to a different file due to inlining.
We need to set the location just before call, or the wrong
location will be on the stack when creating a stack trace.

(In particular, if this happens when evaluating the argument
to an exception's constructor, the function with the "throw"
call won't be on the stack.)

Change-Id: I84106acff5c089fc75e0cf99516f15a14a54886b
Review-Link: https://gwt-review.googlesource.com/#/c/6733/