Update patch set 1
Patch Set 1:

I've been reconsidering this. (Yes, again!) I wonder if we should just use LinkerContext.isOutputCompact(). This is false for "-style pretty" or "-style detailed".

Yes, it's a little strange, but here is the rationale: if anyone wants to read JavaScript, the first thing they will do is set "-style pretty", and then the .js file will magically be readable. This is pretty darn close to "do what I mean." But nobody who cares about performance will launch with it.

Downside: if there is a bug that depends on the installCode setting, it will go away if you turn on "-style pretty". But you can override that setting the "installCode" configuration property directly.

Patch-set: 1