Emit a warning when a JSNI method reference is assigned to a field.
A common error when passing method as callbacks in JSNI occurs for
example in:

  $wnd.aCallback = o.@Class:classback();

where a qualified method reference is set into a field of a different
object without capturing the qualifier. A later call

  $wnd.aCallback();

will bind $wnd to 'this' and the callback; which is probably not desired.

In order to silence this warning the user can explicitly assign the
qualified method reference to a variable as follows.

  var callback =  o.@Class:classback();
  $wnd.aCallback = callback;

Change-Id: I06ae1f9eddc225d3507d556b5fec8faccd7e65e6