Fix CompilationStateBuilder for units that fail JDT compile.
Units that fail JDT compiler are preserved and some passes
rely on those beign in a somewhat consistent state.

On the other hand collecting needed information from these
units might fail while traversing the JDT AST because JDT
doest not guarantee a consistent state for the AST if the
unit has compilation errors.

This patch fixes a breakage when superdevmode recompiles
due to having units with JDT errors incorrectly represented.

TODO: have a saner approach to units that fail to compile
under JDT.

Change-Id: I8c9d7a59d4f4be424b81ba1a46993924e9ffc126