Restore support for derived DataResource in @url CSS extensions
GWT supports references to Data and Image resource using the @url
CSS extension. In 2.6.0, a test was introduced to check that the
type of the method referenced in the @url extension is indeed of
type DataResource or ImageResource. This test is too strict though
and does not allow DataResource and ImageResource subclasses, which
were allowed before 2.6.0. This patch makes the test more lenient
to accept derived DataResource and ImageResource.

Bug: issue 8564
Change-Id: If81e5f1250e28a5a3150b26d3ffa290d780a26a0