Update patch set 5
Patch Set 5:

I finally found some time to update this:

This is now tested on:

 * Win7 IE9
 * Win7 IE10
 * Win7 Safari 5
 * Win7 FF27
 * Win7 Chrome 33
 * Android 4.4 Chrome
 * iOS 7.1 mobile Safari
 * Mac OS Mavericks Safari 7.0.2

I will still be testing this on older Androids and IE11, but this should be fine.


I also wrote a small app to test bookmarking on all browser to make sure that despite using pushState and some browser not correctly displaying the URL that it gets bookmarked properly.
This does not seem to be a problem anymore in latest Safari 7. URL is updated properly in the URL bar and bookmarking works fine. I am not sure whether we want to investigate for Safari 6 seperatly since Safari 5 is fine.

@Goktug
About the three different HistoryImpl classes. I know that we discussed only using two here, but revisiting this I feel better having three implementations:

 * One future implementation using pushState which should become standard for all browsers over time.
 * One implementation for browser that do not support pushstate right now, but support window.addEventListener for hashchange events
 * One implementation for IE8 that does not support window.addEventListener for hashchange events (which we will delete once we drop IE8 with GWT 3.0)

This will leave us (after dropping IE8) support with two implementations, that are both standards based.

Patch-set: 5