CrossSiteIFrameLinker: change installcode default for pretty mode
When using -style pretty or -style detailed, change the default
for "installcode" to false so that the cache.js files are easier
to read.

However, don't change this for optimized builds because it might
have a bad effect on latency for some apps; it's hard to tell
without doing extensive testing.

Most apps shouldn't care about this, but where it matters, the
default can be explicitly overridden:

  <set-configuration-property name="installCode" value="true" />

Change-Id: I98a15ed8c47362c014680e6cddfdab9c1755e67a