Update patch set 3
Patch Set 3: Code-Review-1

> This was already complex and now having new buckets makes it even
 > more complex (systemClasses, serverClasses, webAppClasses,
 > systemClassFromWebAppFirst, allowedFromSystemClassLoader).

Yes, but I really think we should rewrite this from scratch and change the behavior in the next release (or the one after). The complexity comes from the behavior we wanted to have, but will never really achieve.

I expect this change to be "temporary".

 > This might be too naive but could we has similar effect by just
 > changing the patterns via WebAppContext#{set/add}ServerClasses/SystemClasses()?

Unfortunately no. Not without changing the whole behavior of classloading. It'll be needed, but I don't really want to do it in 2.6.1. To begin with, we'd first have to define what behavior we want.

We could setSystemClasses to the same values minus "javax." and with the addition of "javax.servlet." instead of the systemClassFromWebAppFirst bucket, but that only solves half the problem: we want to allow "falling back" to the system classloader for some selected classes (allowedFromSystemClassLoader). This is where Jetty expects the standard classloader delegation to take place (from webapp to system) but we chose that the parent classloader for the webapp would only have bootstrap classes (because most of the time, what's in WEB-INF/lib is also on the system classloader, and a) we want to issue a warning when something should be in WEB-INF/lib but is only found in the system classloader and b) loading those from the system classloader could break things when those classes load resources from the classloader and the user expect them to find things from WEB-INF; e.g. Spring and logging configuration files).

 > Or could we do any other simplification by changing priorities of
 > class loaders?

Neither. The "system classloader" (with gwt-dev.jar) is not in the classloader hierarchy of the webapp, so it's not a matter of priorities.

BTW, I just found a bug in the current patchset.

Patch-set: 3
Label: Code-Review=-1