Update patch set 1
Patch Set 1:

I was thinking about this and I think it would be simpler to rewrite instance method calls rather than their static versions. If this pass is run before MakeCallsStatic then we can guarantee that oportunities are not lost to MakeCallsStatic.

Having at instance methods makes it (1) easier to rescue at CFA (just rescue when the annotated method is rescued), (2) simpler rewrite code (no need to add extra parameter, like in MethodCallTightener) 

What are your thoughts on this?

Patch-set: 1