Update patch set 1
Patch Set 1:

> Any specific region you chose to make UserAgent no longer inherit
 > CoreWithUserAgent, given that that's what all those other modules
 > are using?

I'm basically avoiding circular deps and building reasonable separately compilable pieces.
Another option (without a circular dep) would be making UserAgent a library inheriting the CoreWithUserAgent and keep CoreWithUserAgent a fileset.
That makes browser specific Core and Emulation code to be bundled as part of UserAgent which I think is counter intuitive. I think browser permutation support is a fundamental piece (that's why Core is depending on it) and should be considered as part of Core. The current model I'm suggesting here is having two Cores. One with browser permutations and another one without the permutations. I think easy to understand.

 > I also suspect third-party modules to depend on Core and
 > UserAgent rather than CoreWithUserAgent directly.

I would expected most third_party modules to depend directly on User which pulls in both. For others, if they are not opting into separate compilation, they wouldn't be effected. If they are opting into it, then they would probably need to fix their inherits anyway.

In a follow up patch, I'll investigate separating UserAgent target into 2 pieces; browser permutation support (that will be continued to be inherited from Core as in here) and the UserAgent API which could be a separate lib (i.e. non-fileset module) and can inherit the CoreWithUserAgent. That will make the change fully backward compatible.

 > 
 > I also liked EmulationWithUserAgent, even though it'll go away soon
 > anyway as the only rebinding is for IE6/7.

This only provides a rebinding for code in Core and I think that is better placed next to it. But regardless as you said, this will go away.

Patch-set: 1