Update patch set 1
Patch Set 1:

> > > Any specific region you chose to make UserAgent no longer
 > inherit
 > > > CoreWithUserAgent, given that that's what all those other
 > modules
 > > > are using?
 > >
 > > I'm basically avoiding circular deps and building reasonable
 > > separately compilable pieces.
 > > Another option (without a circular dep) would be making UserAgent
 > a
 > > library inheriting the CoreWithUserAgent and keep
 > CoreWithUserAgent
 > > a fileset.
 > > That makes browser specific Core and Emulation code to be bundled
 > > as part of UserAgent which I think is counter intuitive. I think
 > 
 > I don't think it is counter intuitive. Everything depends on Core.
 > if someone needs UserAgent, he can import Core and UserAgent.
 > 
 > I am now leaning more into not having CoreWithUserAgent nor
 > EmulationWithUserAgent....

The probably means we are thinking differently :)

gwt.user.UserAgent looks like a regular target and there is no clear indication what it would provide. Developers who get used to SDK's will simply think that it provides access to UserAgent API. Any other major behavior difference in the SDK would be a big surprise.

And even if somebody ends up developing a mental model that suggests pulling in UserAgent pulls in browser specific code, then it will be not correct because it will just pull in the specializations for some parts of the SDK (ie. java.lang emul and gwt.Core).

There is too much magic all around and hard to communicate.

On the other hand, it is very simple choice to use the either the GWT Core or GWT Core with useragent support based on what you need in the app. No magic or hard to document behavior involved; GWT Core already known to pull in basic and the other one pulls in the basics with browser specializations.

Hopefully after having Core and CoreWithUserAgents, we can eventually take the necessary step to get rid of CoreWithUserAgent probably in favor of a flag but such a change is more involved and I want to defer that.

Patch-set: 1