Update patch set 2
Patch Set 2:

> > > Couldn't we replace c.g.g.xhr.client.XHR in HttpThrowableReporter
 > > > with JSNI, to effectively drop the dependency?
 > >
 > > I don't think it would worth the trouble. This code doesn't have
 > > server side support and we already have similar functionality
 > with
 > > logging system which also includes components for server-side.
 > 
 > So you're proposing that we defer I5045718051640a4ff9edb83acb0f22ab00bb5a2b
 > until we remove HttpThrowableReporter?
 > (which could happen soon, but still)
 > 
 > Using JSNI here seems straightforward and we wouldn't be in a
 > situation where we deprecate an API in a point release to drop
 > completely in the next release a couple months later (unless you'd
 > propose we wait until after GWT 2.7, I doubt it)

Actually, I was preparing the patch to remove it :)

Especially for this case, I think it is OK to deprecate in 2.6.1 and remove it at 2.7. The class is very unlikely used by developers and missing functionality is quite easy to replace. As some data point this class is not used by any Google app.

On the other hand, if you guys are uncomfortable, I can replace it with JSNI but I really don't think it would worth the effort for this one.

Patch-set: 2