Update patch set 2
Patch Set 2:

> > > > > > Couldn't we replace c.g.g.xhr.client.XHR in
 > > HttpThrowableReporter
 > > > > > > with JSNI, to effectively drop the dependency?
 > > > > >
 > > > > > I don't think it would worth the trouble. This code doesn't
 > > > have
 > > > > > server side support and we already have similar
 > functionality
 > > > > with
 > > > > > logging system which also includes components for
 > > server-side.
 > > > >
 > > > > So you're proposing that we defer I5045718051640a4ff9edb83acb0f22ab00bb5a2b
 > > > > until we remove HttpThrowableReporter?
 > > > > (which could happen soon, but still)
 > > > >
 > > > > Using JSNI here seems straightforward and we wouldn't be in a
 > > > > situation where we deprecate an API in a point release to
 > drop
 > > > > completely in the next release a couple months later (unless
 > > > you'd
 > > > > propose we wait until after GWT 2.7, I doubt it)
 > > >
 > > > Actually, I was preparing the patch to remove it :)
 > > >
 > > > Especially for this case, I think it is OK to deprecate in
 > 2.6.1
 > > > and remove it at 2.7. The class is very unlikely used by
 > > developers
 > > > and missing functionality is quite easy to replace. As some
 > data
 > > > point this class is not used by any Google app.
 > > >
 > > > On the other hand, if you guys are uncomfortable, I can replace
 > > it
 > > > with JSNI but I really don't think it would worth the effort
 > for
 > > > this one.
 > >
 > > Well, I wouldn't want it to set a precedent, and I'm worried that
 > > people would see GWT APIs as unstable and changing, and fear that
 > > they could change under their feet in the future, making them
 > > uncomfortable to use GWT at all.
 > >
 > > Other than that, and for this specific change, I indeed don't
 > worry
 > > much about breaking some people, I'm pretty sure it wouldn't
 > break
 > > anyone ;-)
 > 
 > This code has been there for years untouched, that is the opposite
 > of unstable.. :P

I meant that from the outside, the class is deprecated in a point release and removed a couple months later in the next release. That doesn't leave much time for a user to update his code, let alone if it's not his code but a third-party dependency that uses the class.

As I said, in this specific case, I don't think we'd break much people (if anyone at all), but I wouldn't want to send a signal that could be ill perceived that APIs come and go without giving you the time to adapt (i.e. unstable) and that, as a result, GWT is maybe not something you should rely on.

I still believe we should settle on a clear deprecation policy and follow it to the letter (unless there's a really strong reason to do otherwise).
My policy would be: avoid deprecating things in point releases, and definitely don't remove APIs in point releases; when something is deprecated in one release, it'll likely be gone in the next release (for people running off trunk, that would mean that the deprecation notice is nearly 6 months, but we could also just set a strong minimum at 3 months for example).
I'll propose it to the Steering Committee, as I think it's high time that we settle on something.

 > Anyway, I sent https://gwt-review.googlesource.com/7140 for review
 > that removes the code. I'll leave you to make the final call :)

I'll do the counter-proposal with JSNI then ;-)

Patch-set: 2