Update patch set 2
Patch Set 2:

Uploaded https://gwt-review.googlesource.com/7143

For the deprecation, I don't think one size fits all approach will work. Every change deprecation has its own impact and some might require longer deprecation process and some may not.

For example; there is no reason to introduce an overhead when we know something is not used where we can take a shortcut. Even with the point release deprecation, there will be at least a perceived 2-3 months time-frame between deprecation and deletion of the API which I think is quite fair. And there is no way deletion of an unused API with this level of warning to cause an impression of being 'unstable'.

And keeping API minimum 3 month on the trunk part is also sounds unnecessary to me given that most people already know what it means to run on the trunk.

Also keep in mind, as a special case for the 3.0 release, we may not even be able to give a warning before removal of some stuff as we might miss the 2.7 release window for some deprecations. Being said that we should do our best stick to the standard deprecation process.

Patch-set: 2