Update patch set 2
Patch Set 2:

I think we should stick with deprecating API's for at least one release. How much longer than that seems like a judgement call.

If someone wants to contribute a patch to keep it working and it doesn't impact anything else, I think we can keep it for a while.

Patch-set: 2