Remove source map ranges for most non-statement nodes
The only expression ranges we keep are for inlined Java
functions (for better stack traces) and expressions in loop
statements (to make sure a breakpoint can be set there.)
For other statement types, it seems sufficient to set a
breakpoint at the beginning of the statement.

The main improvement is for debugger usability since we now
disallow setting a breakpoint in the middle of a Java statement
that's been split across multiple lines.

For single-line Java statements, we were already consolidating
ranges in SourceMapRecorder, so there is no change.

Sourcemap size is reduced by a minor amount (about 9%
uncompressed).

(Also removed a few unnecessary overrides.)

Change-Id: Ic108b396260e6c549c893f0b0b5bd83e101ac3af