Update patch set 4
Patch Set 4:

Ooops I'm so used to working on projects that use maven or graddle that I forget that GWT is still using ant.

But I think that we should exclude OS related files. I don't think that every contributor has a .gitignore_global file and so it's better to fix that in the .gitignore of the project directly (and it doesn't cost anything)

Patch-set: 4
Label: Verified=0