Update patch set 3
Patch Set 3:

> So, are we changing so we don't expect svn tools to have all the
 > libraries needed to run this?  If so, then this is fine --
 > otherwise, I think we are better keeping the patches in the same
 > place where we provide patched jars.

Actually this is not a patched jar but a patched data. I think it is
useful to have this stuff in gerrit to see easily changes in each version. 
I dont know what is happening with GWT_TOOLS, but it seems 
it's hard to update since afaik, only googlers can  update it,
maybe it is a subject for the SC.

Patch-set: 3