Update patch set 4
Patch Set 4:

> We should update the documentation with this feature when it is
 > released, maybe a comment in the issue to remember.

Hmm, not sure what to write though. This was something that wasn't supported (and not documented specifically as not being supported), and now is.
People who tried it probably found (or were directed to) the issue in the tracker, so if they starred it they'll be notified that it's fixed.

(on my side, it means I'll be able to use a vanilla GWT as of 2.7 rather than a patched GWT ;-) )

I'll add a "release notes" label to the issue so we can remember it's noteworthy when it's time to release GWT 2.7.

Patch-set: 4