Update patch set 5
Patch Set 5:

> nice env LC_ALL=en_US.utf8 TZ=America/Los_Angeles ant test
 > 

I managed to run tests and there are many of them broken, my question
now is:

most tests are clearly  changes in cldr data like:

    expected: <Total is US$11,305.01>, actual: <Total is $11,305.01>  
    expected: <It is 12:01 on Saturday, 2007 December 01>, actual: <It is 12:01 on 2007 December 1, Saturday>

So changing the expected string should be Ok. 

But other tests like this, are OK to change the expected string as
well or mean that is something going on?

    expected  'The values are 1 and 2>, actual: <The values are 1, 2>

Patch-set: 5
Label: Verified=0