Super Dev Mode UI improvements
- Added the status for each module (whether Super Dev Mode is
on or off). It can be clicked to toggle it on and off. (This
can be useful when compiles are slow, but is also a bit dangerous
because the permutation might be the wrong one for the current
browser.)

- Show any module error message underneath it, instead of as a
tooltip. (They were commonly overlooked before.)

- Changed the footer to show the code server's hostname and port,
and added a tooltip for it.

- Made sure all text is black, regardless of the page's CSS.

- Fixed a problem with the recompile request getting cached on IE9.

Change-Id: If4c1610b251d2c3d5794a7073a5abb38d3ca6123
Review-Link: https://gwt-review.googlesource.com/#/c/7430/