Update patch set 4
Patch Set 4:

(1 comment)

About the compile bookmarklet:

I know it appears as a tip on button, but I rarely come across a developer who knows about this.
I think the main reason why people are not discovering this is, after hitting the button, the button disappears immediately so there is no time for tip to show up.

Patch-set: 4
Label: Verified=0