Update patch set 4
Patch Set 4:

> I know it appears as a tip on button, but I rarely come across a
 > developer who knows about this.
 > I think the main reason why people are not discovering this is,
 > after hitting the button, the button disappears immediately so
 > there is no time for tip to show up.

Yes, this is true. However, I'm not sure how to fix it without cluttering the UI too much. Maybe add a help screen and a (?) that pops it up?

Anyway, they can still recompile if they don't know the trick, so I'm going to defer it to another patch.

Patch-set: 4