Update patch set 1
Patch Set 1:

(4 comments)

I did a quick pass. One of the major question is how JsType would interaction with parent members. Making it automatically expose members from parent is fragile due to overloads. Also the name JsType sounds problematic.

Anyway after thinking and discussing these issues with Roberto; I think we have a different proposal for the all the JsInterop stuff :)

Basically we think that we can support overloading with runtime dispatching (FYI, Nashorn is doing it as well). I was opposing that before because of possible incompatibilities in static vs. runtime dispatching but I think we've figured out a way that would behave in a way that wouldn't be surprising for developers. With that we think we can support any Types to be easily used from javascript side which can change how we structure our problem and the solution.

I can already picture what your reaction will be - "this what Bach was doing" :)

Anyway, let's talk tomorrow.

Patch-set: 1
Label: Verified=0