Refactors out function name extractor in Collector base.
The base implementation Collector.extractName used in
different Collectors and for different purposes.

First use is to get the function name from stack trace in
Mozilla browsers.

Second use is to get the function name out of toString call.
This part is used to extract function name in CollectorLegacy
and CollectorEmul which have access to function objects.
In many browsers function.name is already provided and doesn't
need parsing. If it is not available, it needs to be parsed only
once and put into the name field for future references.

The patch introduces a specialized extract name for the second
use case and optimizes for that case. The existing implementation
is also moved out of the based classed into CollectorMozilla.


Change-Id: I91d863ebad299650252a7ce2bf7b74bda40ab981