Export user.agent and locale to __gwt_activeModules more reliably.
Embed these properties into the initial fragment (in the $permProps
global variable) and export them when the module has finished
loading. This should work even when the developer has restricted
the generated permutations or they are using server-side script
selection (SSSS).

The properties are embedded for all linkers, but for now they are
only added to __gwt_activeModules for CrossSiteIframeLinker.

The js.embedded.properties config variable can be used to add
more properties to the list.

Change-Id: I919c43d0ac074d8e21f404ef1f7339e9e7f1acba