Export user.agent and locale more reliably.
The Super Dev Mode bookmarklet calls
__gwt_activeModules[(module name)].bindings() to find out
which bindings it should compile with. However, this isn't
reliable when the developer restricts permutations by
setting these variables to a fixed value or when using
server side script selection. This causes Super Dev Mode
to compile more permutations than necessary.

To fix, embed these properties into the initial fragment
(in the $permProps variable) and export them when the
module has finished loading.

The initial fragment always has $permProps, but for now
it's only used in CrossSiteIframeLinker.

The js.embedded.properties config variable can be used to
add more properties to the list.

Bug: issue 7458
Change-Id: I2c7bf4635139eef3d938c4bd8aaaa19499c43bc3