Update patch set 1
Patch Set 1:

> I kind of instinctively prefer Option 1 so we preserve the ability
 > to optimize away instanceof operations, since it is against my
 > nature to like changes which may increase code size, but I do
 > accept that instanceof on native arrays happens almost never, and
 > well, this simplifies the compiler.

I tried option 1 (and it is in the first patch set of this review), and indeed it fixes this problem, but I am not sure it is completely correct. I tested the current cl in CM and the uncompressed output only changed a few bytes.

Patch-set: 1
Label: Verified=0