Update patch set 1
Patch Set 1: Code-Review-1

This is actually a problem on StackTraceCreator as it fails to parse the stack trace properly for IE and it is better to be fixed there.

I guess this problem occurs with IE11?

Patch-set: 1
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1