Update patch set 1
Patch Set 1:

Let's first create a bug report with an example stack trace so we could track this in issue tracker.

Then next step would be to reproduce the problem in a test case. I think I have already covered that in StackTraceCreatorCollectorTest#_disabled_testIE_10 but if it is not covered, we should generate an example of the offending stack trace.

Also, if you don't mind, it would be good to wait a few weeks before the patch as I'm actively changing these code.

Patch-set: 1