Update patch set 1
Patch Set 1:

As it commented in build.xml Jetty is repackaged for Google
internl usage, and not repackaged for normal ditribution.
So it doesn't compile out of the box; to fix it, the package

  com.google.gwt.thirdparty.org.eclipse.jetty

must be replaced with

  org.eclipse.jetty

in WebServer.java. Another approach to fix it would be to put
repackaged Jetty under

  GWT_TOOLS/lib/jetty/jetty-<VER>/jetty-VER-repackaged.jar

Patch-set: 1
Label: Verified=0