Update patch set 1
Patch Set 1:

> As it commented in build.xml Jetty is repackaged for Google
 > internl usage, and not repackaged for normal ditribution.
 > So it doesn't compile out of the box; to fix it, the package
 > 
 >  com.google.gwt.thirdparty.org.eclipse.jetty
 > 
 > must be replaced with
 > 
 > org.eclipse.jetty
 > 
 > in WebServer.java. Another approach to fix it would be to put
 > repackaged Jetty under
 > 
 >  GWT_TOOLS/lib/jetty/jetty-<VER>/jetty-VER-repackaged.jar

Actually, I think we should move the code to either using a repackaged Jetty everywhere or not using a repackaged Jetty at all (i.e. update dev/codeserver). I lean towards the second option, as we're looking at modularizing GWT in the future, and using true dependencies rather than fat-jars (at least on the OSS side of the project; if fat-jars are easier to manage for Google, they have their own build process anyway).

AFAIK, the way it's done here was to make it easier to sync the OSS repo with Google's internal repo (i.e. no need to mangle the imports). I think we should instead make it easier for the OSS community to work with the code and contribute; the package-mangling is done for dev/core and user/ already, and AFAICT it's a one-time setup for an automated tool.

Patch-set: 1