Update patch set 1
Patch Set 1:

Hi, thanks for patch. I don't have time for this today, but I'm going to patch this in and get our build system working; we can handle this the way we do all the other GWT code that uses repackaged dependencies internally.

(We do need to repackage it since there are multiple versions of Jetty in use and we don't want to pick up the wrong one.)

Patch-set: 1
Label: Code-Review=0