Chrome document scrolling behavior depends now on the doctype.
Recently, Chrome change its behavior concerning the document scrolling
and follow the standard behavior. In strict mode, the document scrolling
should be applied to document.documentElement but in quirks mode it
should be applied to document.body.
Safari always applies document scrolling to the body element even in
strict mode.

See fixed chromium issue:
https://code.google.com/p/chromium/issues/detail?id=303131

JsFiddle showing the issue: http://jsfiddle.net/b6j6y/7/

Note: This change of behavior of chrome breaks codes that are listening
on window scroll (Window.addWindowScrollHandler) because the the
ScrollEvent instance always returns 0 for scrollTop and scrollLeft on
chrome. This is fixed with this patch.

Change-Id: Ibad54ccb8bc36102e2be7e69f2e08495932ca620