Update patch set 3
Patch Set 3:

(1 comment)

One last thought: `set-property-fallback` seems to have been tailor-made for locale settings, so that there is a reasonable default to lean on. The goal of this patch is to bring that feature to user.agent as well, but without any functionality changes until a user asks for it.

Patch-set: 3