Update patch set 3
Patch Set 3:

> Can devmode handler be also deprecated in favour of console handler
 > by adding
 > 
 > 
 > if (!GWT.isScript()) GWT.log(msg);
 > 
 > 
 > to the console handler?

Interesting idea.

There is a semantic difference as GWT.log can be bound to different things and disabled in production. I think the decision should be based on the future of GWT.log instead. (e.g. I will enable logging for devmode by default so after that point are we going to still keep GWT.log?).

Patch-set: 3